requirements.txt
py2neo>=2021.2.4

# Knowledge Fusion Module Dependencies
sentence-transformers>=2.2.2
langchain>=0.1.0
langchain-openai>=0.0.5
pydantic>=2.5.0
numpy>=1.24.0
python-dotenv>=1.0.0
tf-keras>=2.20.0  # Required for sentence-transformers compatibility
